Revert "Asio+Beast-based endpoint server (#274)"
This reverts commit d686b8eb4c70ea3fa20a32d6f314f17efa941b13.

+++ src/transport/MessageEndpoint.cpp
@@ -122,9 +122,8 @@ zmq::socket_t socketFactory(zmq::socket_type socketType,
                     break;
                 }
                 default: {
-                    SPDLOG_ERROR("Invalid bind socket type {} ({})",
-                                 static_cast<int>(socketType),
-                                 address);
+                    SPDLOG_ERROR(
+                      "Invalid bind socket type {} ({})", socketType, address);
                     throw std::runtime_error(
                       "Binding with invalid socket type");
                 }
@@ -177,7 +176,7 @@ zmq::socket_t socketFactory(zmq::socket_type socketType,
                 }
                 default: {
                     SPDLOG_ERROR("Invalid connect socket type {} ({})",
-                                 static_cast<int>(socketType),
+                                 socketType,
                                  address);
                     throw std::runtime_error(
                       "Connecting with unrecognized socket type");
